moved cif2astr() structure_analyzer.py -> io/cif.py
